store.js 파일 분리 -> store.js & validateData.js
